src/requirements.txt
fastapi
fasttext-wheel
nltk
python-box[yaml]
sacremoses
sentencepiece
torch
transformers
uvicorn[standard]
